demo/requirements.txt
streamlit>=1.24.0
langchain>=0.0.217
openai
python-dotenv